+++ backend/config/urls.py
@@ -44,6 +44,7 @@ urlpatterns = [
     path('api/', include('galleries.urls')),
     path('api/', include('coauthoring.urls')),
     path('api/', include('concepts.urls')),
+    path('api/', include('venues.urls')),
 ]
 
 if settings.DEBUG: